% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{calculate_dist_samples}
\alias{calculate_dist_samples}
\title{Calculate Distances Between Transformed Coordinates}
\usage{
calculate_dist_samples(values, coord1, coord2)
}
\arguments{
\item{values}{Numeric vector. Alignment parameters obtained from structure
alignment.
 The vector includes the following elements:\cr
 - dx: Translation along the x-axis.\cr
 - dy: Translation along the y-axis.\cr
 - dz: Translation along the z-axis.\cr
 - theta: Rotation angle around the x-axis.\cr
 - phi: Rotation angle around the y-axis.\cr
 - psi: Rotation angle around the z-axis.}

\item{coord1}{Matrix. 3D coordinates of the first structure's atoms
(rows: dimensions, columns: atoms). This parameter should be a 4xN matrix,
where N is the number of atoms. The matrix is structured such that each
column corresponds to the 3D coordinates of an alpha-carbon atom in the
first structure, and each row represents a dimension (x, y, z, and a
placeholder coordinate). The placeholder coordinate is set to 1 for each
atom, facilitating matrix transformations.}

\item{coord2}{Matrix. 3D coordinates of the second structure's atoms
(rows: dimensions, columns: atoms). This parameter should be a 4xN matrix,
where N is the number of atoms. The matrix is structured such that each
column corresponds to the 3D coordinates of an alpha-carbon atom in the
second structure, and each row represents a dimension (x, y, z, and a
placeholder coordinate). The placeholder coordinate is set to 1 for each
atom, facilitating matrix transformations.}
}
\value{
A matrix of Euclidean distances between the transformed coordinates
 (coord) and the original coordinates of the first structure (coord1).
}
\description{
This function calculates the Euclidean distances between transformed
coordinates based on a 4x4 transformation matrix obtained from the alignment
parameters. The alignment parameters include translation along x, y, z axes
(dx, dy, dz), and rotation angles (theta, phi, psi). The transformed
coordinates are calculated by applying the transformation matrix to the
original coordinates of the second structure (coord2).
}
\examples{
\dontrun{
# Example: Calculate distances between transformed coordinates
alignment_params <- c(dx = 1.0, dy = 2.0, dz = 0.5, theta = 0.1, phi = 0.2,
                      psi = 0.3
                      )
original_coordinates <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8), nrow = 4,
                               byrow = TRUE
                               )
transformed_coordinates <- matrix(c(2, 3, 4, 5, 6, 7, 1, 8), nrow = 4,
                                  byrow = TRUE
                                  )
distances <- calculate_dist_samples(alignment_params, original_coordinates,
                                    transformed_coordinates
                                   )
}

}
\references{
Jordi Cruzado (https://math.stackexchange.com/users/96872/jordi-cruzado),
Explain 3d transformation matrix..., URL (version: 2023-05-30):
\href{https://math.stackexchange.com/q/532974}{Link}
}
\seealso{
\code{\link{get_matrix}} for obtaining the transformation matrix from
  alignment parameters.
}
