% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{calculate_rmsd}
\alias{calculate_rmsd}
\title{Calculate Root Mean Square Deviation (RMSD) Between Transformed Coordinates}
\usage{
calculate_rmsd(values, coord1, coord2)
}
\arguments{
\item{values}{Numeric vector. Alignment parameters obtained from structure
alignment.
 The vector includes the following elements:\cr
 - dx: Translation along the x-axis.\cr
 - dy: Translation along the y-axis.\cr
 - dz: Translation along the z-axis.\cr
 - theta: Rotation angle around the x-axis.\cr
 - phi: Rotation angle around the y-axis.\cr
 - psi: Rotation angle around the z-axis.}

\item{coord1}{Matrix. 3D coordinates of the first structure's atoms
(rows: dimensions, columns: atoms).}

\item{coord2}{Matrix. 3D coordinates of the second structure's atoms
(rows: dimensions, columns: atoms).}
}
\value{
The Root Mean Square Deviation (RMSD) between the transformed
coordinates and the original coordinates of the structures.
}
\description{
This function calculates the Root Mean Square Deviation (RMSD) between the
transformed coordinates and the original coordinates of two structures. The
RMSD is a measure of the average distance between corresponding atoms in the
two structures after optimal superposition. The alignment parameters,
original coordinates of the first structure (coord1), and original
coordinates of the second structure (coord2) are required. Note that unlike
\code{\link{get_rmsd}}, this function is a helper function that performs
the actual RMSD calculation.
}
\examples{
\dontrun{
# Example: Calculate RMSD between transformed coordinates
alignment_params <- c(dx = 1.0, dy = 2.0, dz = 0.5, theta = 0.1, phi = 0.2,
                      psi = 0.3
                      )
original_coordinates <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8), nrow = 4,
                               byrow = TRUE
                               )
transformed_coordinates <- matrix(c(2, 3, 4, 5, 6, 7, 1, 8), nrow = 4,
                                  byrow = TRUE
                                  )
rmsd_value <- calculate_rmsd(alignment_params, original_coordinates,
                             transformed_coordinates
                             )
}

}
\seealso{
\code{\link{calculate_dist_samples}} for calculating distances between
transformed coordinates.
}
