% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{get_matrix}
\alias{get_matrix}
\title{Get Transformation Matrix from Alignment Parameters}
\usage{
get_matrix(values)
}
\arguments{
\item{values}{Numeric vector. Alignment parameters obtained from structure
alignment.
 The vector includes the following elements:
 - dx: Translation along the x-axis.
 - dy: Translation along the y-axis.
 - dz: Translation along the z-axis.
 - theta: Rotation angle around the x-axis.
 - phi: Rotation angle around the y-axis.
 - psi: Rotation angle around the z-axis.}
}
\value{
A 4x4 transformation matrix for aligning structures based on the
 input alignment parameters.
}
\description{
This function constructs a 4x4 transformation matrix based on the alignment
parameters obtained from the structure alignment. The alignment parameters
include translation along x, y, z axes (dx, dy, dz), and rotation angles
(theta, phi, psi). The resulting transformation matrix can be used to
transform the coordinates of one structure to align with the other.
}
\examples{
\dontrun{
# Example: Get transformation matrix for structure alignment
alignment_params <- c(dx = 1.0, dy = 2.0, dz = 0.5, theta = 0.1, phi = 0.2,
                      psi = 0.3)
transformation_matrix <- get_matrix(alignment_params)
}

}
\references{
Jordi Cruzado (https://math.stackexchange.com/users/96872/jordi-cruzado),
Explain 3d transformation matrix..., URL (version: 2023-05-30):
\href{https://math.stackexchange.com/q/532974}{Link}
}
