% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{get_rmsd}
\alias{get_rmsd}
\title{Calculate Root Mean Square Deviation (RMSD) Between Two Protein Structures}
\usage{
get_rmsd(pdb1, pdb2, chain1 = "A", chain2 = "A", method)
}
\arguments{
\item{pdb1}{Character. Path to the PDB file of the first protein structure.}

\item{pdb2}{Character. Path to the PDB file of the second protein structure.}

\item{chain1}{Character. Chain identifier for the first protein structure (default: 'A').}

\item{chain2}{Character. Chain identifier for the second protein structure (default: 'A').}

\item{method}{Character. The method used for structure alignment, either "alignment" or "index".}
}
\value{
The Root Mean Square Deviation (RMSD) between the two protein structures.
}
\description{
This function calculates the Root Mean Square Deviation (RMSD) between two protein structures
based on their atomic coordinates. The RMSD is a measure of the average distance between
corresponding atoms in the two structures, providing an indication of structural similarity.
}
\examples{
\dontrun{
# Example: Calculate RMSD between two protein structures
pdb_file1 <- "path/to/structure1.pdb"
pdb_file2 <- "path/to/structure2.pdb"
chain_id1 <- 'A'
chain_id2 <- 'B'
alignment_method <- "alignment"
rmsd_value <- get_rmsd(pdb_file1, pdb_file2, chain_id1, chain_id2, alignment_method)
}

}
\seealso{
\code{\link{get_alignment}} for obtaining structural alignment between two protein structures.
}
