% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{get_tm_samples}
\alias{get_tm_samples}
\title{Get TM local scores from Protein Structure Alignment}
\usage{
get_tm_samples(alignment)
}
\arguments{
\item{alignment}{List. Structure alignment results, including alignment
parameters, coordinates, and other information.
The list should contain the following elements:\cr
- N: Numeric. Number of common residues in the alignment. \cr
- coord1: Matrix. 3D coordinates of CA atoms for common residues in the
  first structure.\cr
- coord2: Matrix. 3D coordinates of CA atoms for common residues in the
  second structure.\cr
- values: Numeric vector. Alignment parameters.}
}
\value{
Numeric vector. TM samples of the protein structure alignment.
}
\description{
This function calculates the TM local scores from the alignment parameters
and coordinates obtained in a structural alignment between two protein
structures.
}
\examples{
\dontrun{
# Example: Calculate TM samples from alignment
pdb_file1 <- system.file("extdata", "1LNIA_decoy1_4.pdb",
                          package="TMscoreAlign"
                          )
pdb_file2 <- system.file("extdata", "1LNIA_decoy2_180.pdb",
                          package="TMscoreAlign"
                          )
alignment_results <- get_alignment(pdb_file1, pdb_file2,
                                  chain1 = 'A', chain2 = 'A',
                                  method = "alignment"
                                  )
tm_samples <- get_tm_samples(alignment_results)
print(tm_samples)
}

}
\references{
Zhang, Y., and Skolnick, J. (2004). Scoring function for automated assessment
of protein structure template quality. \emph{Proteins, Structure, Function,
and Bioinformatics}, 57(4), 702–710.
\href{https://doi.org/10.1002/prot.20264}{Link}
}
\seealso{
\code{\link{get_alignment}} for obtaining alignment details.
\code{\link{optimize_alignment}} for optimizing alignment parameters.
\code{\link{estimate_d0}} for estimating initial distance parameters.
\code{\link{tm_samples}} for calculating TM local scores
}
