% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{load_data_alignment}
\alias{load_data_alignment}
\title{Load Data and Perform Sequence Alignment or Index-based Analysis}
\usage{
load_data_alignment(
  pdb_file1,
  pdb_file2,
  chain1 = "A",
  chain2 = "A",
  method = "alignment"
)
}
\arguments{
\item{pdb_file1}{Character string. Path to the first PDB file.}

\item{pdb_file2}{Character string. Path to the second PDB file.}

\item{chain1}{Character string. Chain identifier for the first PDB file
(default is 'A').}

\item{chain2}{Character string. Chain identifier for the second PDB file
(default is 'A').}

\item{method}{Character string. Method for residue selection.
Options: "alignment" or "index"\cr
 - "alignment": Perform sequence alignment using pairwiseAlignment from
    Bioconductor.\cr
 - "index": Use residue indices based on the specified chains to identify
    common residues.}
}
\value{
A list containing the following components:\cr
  - coord1: 3D coordinates (matrix) of CA atoms for common residues in the
     first structure.\cr
  - coord2: 3D coordinates (matrix) of CA atoms for common residues in the
     second structure.\cr
  - N: Number of common residues.
}
\description{
This function reads two Protein Data Bank (PDB) files, extracts the CA atom
coordinates for common residues between the specified chains, and returns the
coordinates along with the number of common residues.
}
\examples{
\dontrun{
# Example 1: Perform sequence alignment
pdb_file1 <- system.file("extdata", "1LNIA_decoy1_4.pdb",
                          package="TMscoreAlign"
                          )
pdb_file2 <- system.file("extdata", "1LNIA_decoy2_180.pdb",
                          package="TMscoreAlign"
                          )
result_alignment <- load_data_alignment(pdb_file1,
                                        pdb_file2,
                                        method = "alignment"
                                        )

# Example 2: Use residue indices for common residue selection
result_index <- load_data_alignment(pdb_file1,
                                    pdb_file2,
                                    method = "index")
}

}
\references{
Grant, B.J. et al. (2006). Bio3D: An R package for the comparative analysis
of protein structures. \emph{Bioinformatics}, 22, 2695--2696.
\href{https://doi.org/10.1093/bioinformatics/btl461}{Link}, R package version
2.4.4, \href{https://cran.r-project.org/web/packages/bio3d/index.html}{Link}

Pagès H, Aboyoun P, Gentleman R, DebRoy S (2023). Biostrings: Efficient
manipulation of biological strings.
\href{https://doi.org/10.18129/B9.bioc.Biostrings}{Link}, R package version
2.68.1, \href{https://bioconductor.org/packages/Biostrings}{Link}
}
\seealso{
\code{\link{pairwiseAlignment}} for sequence alignment.
\code{\link{atom.select}} for atom selection in the \code{\link{bio3d}}
  package.
}
