% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{optimize}
\alias{optimize}
\title{Optimize Protein Structure Alignment Parameters}
\usage{
optimize(alignment, restart = TRUE)
}
\arguments{
\item{alignment}{List. Structure alignment results, including alignment
parameters, coordinates, and other information.
The list should contain the following elements:
- N: Numeric. Number of common residues in the alignment.
- coord1: Matrix. 3D coordinates of CA atoms for common residues in the
  first structure.
- coord2: Matrix. 3D coordinates of CA atoms for common residues in the
  second structure.
- values: Numeric vector. Initial alignment parameters.}

\item{restart}{Logical. If TRUE, the optimization starts from the default
values. If FALSE, it continues from the current alignment parameters.}
}
\value{
The input alignment list updated with optimized parameters
}
\description{
This function optimizes the alignment parameters for better accuracy of the
structural alignment between two protein structures.
}
\examples{
\dontrun{
# Example: Optimize alignment parameters
alignment_results <- get_alignment("structure1.pdb", "structure2.pdb",
                                  chain1 = 'A', chain2 = 'A',
                                  method = "alignment", optimize = FALSE)
optimized_results <- optimize(alignment_results, restart = TRUE)
print(optimized_results)
}

}
\references{
Byrd, R. H., Lu, P., Nocedal, J., & Zhu, C. (1995).
A limited memory algorithm for bound constrained optimization.
\emph{SIAM Journal on Scientific Computing}, 16(5), 1190–1208.
\href{https://epubs.siam.org/doi/10.1137/0916069}{Link}
}
\seealso{
\code{\link{get_default_values}} for parameter initialization.
\code{\link{tm}} and \code{\link{rmsd}} for TM-score and RMSD calculation.
\code{\link{optim}} for optimization functions.
}
