% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{optimize_alignment}
\alias{optimize_alignment}
\title{Optimize Protein Structure Alignment Parameters}
\usage{
optimize_alignment(alignment, restart = FALSE, maxit = 300)
}
\arguments{
\item{alignment}{List. Structure alignment results, including alignment
parameters, coordinates, and other information.
The list should contain the following elements:\cr
- N: Integer Number of common residues in the alignment.\cr
- coord1: Matrix. 3D coordinates of CA atoms for common residues in the
  first structure.\cr
- coord2: Matrix. 3D coordinates of CA atoms for common residues in the
  second structure.\cr
- values: Numeric vector. Initial alignment parameters.}

\item{restart}{Logical. If TRUE, the optimization starts from the default
values. If FALSE, it continues from the current alignment parameters.}

\item{maxit}{Numeric. The maximum number of iterations for optimization.}
}
\value{
The input alignment list updated with optimized parameters
}
\description{
This function optimizes the alignment parameters for better accuracy of the
structural alignment between two protein structures. Optimization is
performed via the limited-memory Broyden–Fletcher–Goldfarb–Shanno algorithm.
}
\examples{
\dontrun{
# Example: Optimize alignment parameters
pdb_file1 <- system.file("extdata", "1LNIA_decoy1_4.pdb",
                          package="TMscoreAlign")
pdb_file2 <- system.file("extdata", "1LNIA_decoy2_180.pdb",
                          package="TMscoreAlign")
alignment_results <- get_alignment(pdb_file1, pdb_file2,
                                  chain1 = 'A', chain2 = 'A',
                                  method = "alignment", optimize = FALSE)
optimized_results <- optimize_alignment(alignment_results, restart = TRUE)
print(optimized_results)
}

}
\references{
Byrd, R. H., Lu, P., Nocedal, J., & Zhu, C. (1995).
A limited memory algorithm for bound constrained optimization.
\emph{SIAM Journal on Scientific Computing}, 16(5), 1190–1208.
\href{https://epubs.siam.org/doi/10.1137/0916069}{Link}
}
\seealso{
\code{\link{get_default_values}} for parameter initialization.
\code{\link{optim}} for optimization functions.
\code{\link{calculate_tmscore}} and \code{\link{calculate_rmsd}} for
TM-score and RMSD calculation.
}
