% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_pdb2}
\alias{sample_pdb2}
\title{PDB information for 1LNIA decoy2_180}
\format{
Returns a list of class "pdb" representing the decoy2_180 of the
   1LNIA protein. This list has the following components as described by
   the bio3d documentation:
\describe{
 \item{atom}{a data.frame containing all atomic coordinate ATOM and HETATM
     data, with a row per ATOM/HETATM and a column per record type
     }
 \item{helix}{‘start’, ‘end’ and ‘length’ of H type sse, where start and end
     are residue numbers “resno”
     }
 \item{sheet}{‘start’, ‘end’ and ‘length’ of E type sse, where start and end
     are residue numbers “resno”
     }
 \item{seqres}{sequence from SEQRES field.}
 \item{xyz}{a numeric matrix of class "xyz" containing the ATOM and HETATM
     coordinate data
     }
 \item{calpha}{logical vector with length equal to nrow(atom) with TRUE
     values indicating a C-alpha “elety”
     }
 \item{remark}{a list object containing information taken from REMARK
     records of the PDB file
     }
 \item{call}{the matched call}

}
}
\source{
3DRobot_set from Zhang Lab \cr
Deng, H., Jia, Y., and Zhang, Y. (2016). 3DRobot: Automated Generation of
Diverse and Well-packed Protein Structure Decoys. \emph{Boinformatics}, 32,
378-387. \href{https://doi.org/10.1093/bioinformatics/btv601}{Link}
}
\usage{
sample_pdb2
}
\description{
Correctly formatted PDB file. The PDB file contains structural information
about a decoy structure (decoy2_180) that represents a different conformation
of the protein ILNIA. This structure was pre-generated by the 3DRobot program
developed by the Zhang Lab.
}
\examples{
\dontrun{
 sample_pdb2
}
}
\keyword{datasets}
