% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{tm}
\alias{tm}
\title{Calculate Average TM-Score Between Transformed Coordinates}
\usage{
tm(values, coord1, coord2, d02)
}
\arguments{
\item{values}{Numeric vector. Alignment parameters obtained from structure
alignment.
 The vector includes the following elements:
 - dx: Translation along the x-axis.
 - dy: Translation along the y-axis.
 - dz: Translation along the z-axis.
 - theta: Rotation angle around the x-axis.
 - phi: Rotation angle around the y-axis.
 - psi: Rotation angle around the z-axis.}

\item{coord1}{Matrix. 3D coordinates of the first structure's atoms
(rows: dimensions, columns: atoms).}

\item{coord2}{Matrix. 3D coordinates of the second structure's atoms
(rows: dimensions, columns: atoms).}

\item{d02}{Numeric. The square of the d0 parameter, a critical parameter in
the TM-Score calculation.
 It represents a distance cutoff for defining structural similarity.}
}
\value{
The average TM-Score between the transformed coordinates and the
 original coordinates of the structures.
}
\description{
This function calculates the average TM-Score (Template Modeling Score)
between the transformed coordinates and the original coordinates of two
structures. The TM-Score is a measure of structural similarity and is
commonly used in structural bioinformatics. The alignment parameters,
original coordinates of the first structure (coord1), original coordinates of
the second structure (coord2), and the square of the d0 parameter (d02) are
required.
}
\examples{
\dontrun{
# Example: Calculate average TM-Score between transformed coordinates
alignment_params <- c(dx = 1.0, dy = 2.0, dz = 0.5, theta = 0.1, phi = 0.2,
                      psi = 0.3)
original_coordinates <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3, byrow = TRUE)
transformed_coordinates <- matrix(c(2, 3, 4, 5, 6, 7), nrow = 3,
                                  byrow = TRUE)
d0_squared <- 5.0
avg_tm_score <- tm(alignment_params, original_coordinates,
                   transformed_coordinates, d0_squared)
}

}
\references{
Zhang, Y., and Skolnick, J. (2004). Scoring function for automated assessment
of protein structure template quality. \emph{Proteins, Structure, Function,
and Bioinformatics}, 57(4), 702–710.
\href{https://doi.org/10.1002/prot.20264}{Link}
}
\seealso{
\code{\link{tm_samples}} for calculating individual TM-Scores.
}
