% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_alignment.R
\name{visualize_alignment_pdb}
\alias{visualize_alignment_pdb}
\title{Visualize Protein Structure Alignment Using 3Dmol}
\usage{
visualize_alignment_pdb(
  alignment_pdb = "out.pdb",
  chain1 = "#636efa",
  chain2 = "#ff7f0e"
)
}
\arguments{
\item{alignment_pdb}{Character. Path to the PDB file containing the aligned
coordinates. The PDB file should include both the original and
transformed coordinates of the protein structures.}

\item{chain1}{Character. Color code (hexadecimal) for the first protein
structure (default: "#636efa").}

\item{chain2}{Character. Color code (hexadecimal) for the second protein
structure (default: "#ff7f0e").}
}
\value{
A 3Dmol viewer displaying the aligned protein structures.
}
\description{
This function visualizes the structural alignment of two protein structures
by displaying the aligned coordinates in a 3D viewer. The viewer is set up
using 3Dmol, and the alignment is represented by color-coded cartoon-style
structures for each protein chain.
}
\examples{
\dontrun{
# Example: Visualize protein structure alignment
pdb_file1 <- system.file("extdata", "1LNIA_decoy1_4.pdb",
                          package="TMscoreAlign")
pdb_file2 <- system.file("extdata", "1LNIA_decoy2_180.pdb",
                          package="TMscoreAlign")
alignment_results <- get_alignment(pdb_file1, pdb_file2,
                                    chain1 = 'A', chain2 = 'A',
                                    method = "alignment")
write_pdb(alignment_results, outputfile = "aligned_structure.pdb",
         appended = TRUE, pdb1 = pdb_file1, pdb2 = pdb_file2,
         chain_1 = 'A', chain_2 = 'A')
alignment_pdb_file <- "aligned_structure.pdb"
chain1_color <- "#636efa"  # Blue
chain2_color <- "#ff7f0e"  # Orange
visualize_alignment_pdb(alignment_pdb_file, chain1 = chain1_color,
                        chain2 = chain2_color)
}

}
\references{
Su W, Johnston B (2021). r3dmol: Create Interactive 3D Visualizations of
Molecular Data. R package version 0.1.2.
\href{https://CRAN.R-project.org/package=r3dmol}{Link}.
}
\seealso{
\code{\link{write_pdb}} for generating a PDB file with transformed
  coordinates.
\code{\link{get_alignment}} for obtaining structural alignment between two
  protein structures.
}
