% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_alignment.R
\name{write_pdb}
\alias{write_pdb}
\title{Write Transformed Coordinates to PDB File}
\usage{
write_pdb(
  alignment,
  outputfile = "out.pdb",
  appended = TRUE,
  pdb1,
  pdb2,
  chain_1,
  chain_2
)
}
\arguments{
\item{alignment}{List. Structure alignment results, including alignment
parameters and transformed coordinates.
 The list should contain the following elements:
 - coord1: 3D coordinates (matrix) of CA atoms for common residues in the
    first structure.
 - coord2: 3D coordinates (matrix) of CA atoms for common residues in the
    second structure.
 - values: Numeric vector. Alignment parameters obtained from structure
  alignment.
 - rmsd: Numeric. Root Mean Square Deviation (RMSD) between the two protein
  structures.}

\item{outputfile}{Character. The name of the output PDB file to be created
(default: "out.pdb").}

\item{appended}{Logical. If TRUE, the transformed coordinates are appended to
the original coordinates
 in the output PDB file. If FALSE, the output PDB file contains only the
  transformed coordinates.}

\item{pdb1}{Character. Path to the PDB file of the first protein structure.}

\item{pdb2}{Character. Path to the PDB file of the second protein structure.}

\item{chain_1}{Character. Chain identifier for the first protein structure in
output file.}

\item{chain_2}{Character. Chain identifier for the second protein structure
in output file.}
}
\description{
This function writes the transformed coordinates obtained from a structure
alignment to a new PDB file. The transformed coordinates are generated by
applying the alignment parameters to the original coordinates of the second
protein structure. The resulting PDB file includes the transformed
coordinates alongside the original coordinates of the first structure.
}
\examples{
\dontrun{
# Example: Write transformed coordinates to a new PDB file
alignment_results <- get_alignment("structure1.pdb", "structure2.pdb",
                                    chain1 = 'A', chain2 = 'A',
                                    method = "alignment")
write_pdb(alignment_results, outputfile = "aligned_structure.pdb",
         appended = TRUE, pdb1 = "structure1.pdb", pdb2 = "structure2.pdb",
         chain_1 = 'A', chain_2 = 'A')
}

}
\references{
Berman, H., Henrick, K. and Nakamura, H.
Announcing the worldwide Protein Data Bank.
\emph{Nature Structural & Molecular Biology}, 10, 980 (2003).
\href{https://doi.org/10.1038/nsb1203-980}
}
\seealso{
\code{\link{get_alignment}} for obtaining structural alignment between two
  protein structures.
\code{\link{get_matrix}} for obtaining the transformation matrix from
  alignment parameters.
}
