% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{get_alignment}
\alias{get_alignment}
\title{Get Protein Structure Alignment}
\usage{
get_alignment(pdb1, pdb2, chain1 = "A", chain2 = "A", method, optimize = TRUE)
}
\arguments{
\item{pdb1}{Character. Path to the PDB file of the first protein structure.}

\item{pdb2}{Character. Path to the PDB file of the second protein structure.}

\item{chain1}{Character. Chain identifier for the first protein structure.
Defaults to 'A'.}

\item{chain2}{Character. Chain identifier for the second protein structure.
Defaults to 'A'.}

\item{method}{Character. Alignment method to be used. Options are "alignment"
for sequence-based alignment or "index" for index-based alignment.}

\item{optimize}{Logical. If TRUE, the alignment parameters are optimized for
better accuracy. Defaults to TRUE.}
}
\value{
A list containing the following elements: \cr
  - N: Integer. Number of common residues in the alignment. \cr
  - coord1: Matrix. 3D coordinates of CA atoms for common residues in the
    first structure. \cr
  - coord2: Matrix. 3D coordinates of CA atoms for common residues in the
    second structure. \cr
  - values: Double vector. Alignment parameters obtained from structure
    alignment.
}
\description{
This function performs structural alignment between two protein structures
specified by PDB files. It can use either a specified alignment method or
index-based alignment. The alignment parameters, including translation and
rotation values, are initialized and can be further optimized.
}
\examples{
\dontrun{
# Example: Get structural alignment
pdb_file1 <- system.file("extdata", "1LNIA_decoy1_4.pdb",
                          package="TMscoreAlign")
pdb_file2 <- system.file("extdata", "1LNIA_decoy2_180.pdb",
                          package="TMscoreAlign")
alignment_results <- get_alignment(pdb_file1, pdb_file2,
                                  chain1 = 'A', chain2 = 'A',
                                  method = "alignment", optimize = TRUE)
print(alignment_results)
}

}
\seealso{
\code{\link{load_data_alignment}} for loading data for structure alignment.
\code{\link{optimize_alignment}} for optimizing alignment parameters.
}
