% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{get_default_values}
\alias{get_default_values}
\title{Get Default Values for Structure Alignment Parameters}
\usage{
get_default_values(coord1, coord2)
}
\arguments{
\item{coord1}{Matrix. 3D coordinates of the first structure's atoms (rows: dimensions, columns: atoms).}

\item{coord2}{Matrix. 3D coordinates of the second structure's atoms (rows: dimensions, columns: atoms).}
}
\value{
A numeric vector containing default values for the structure alignment parameters:
  - dx: Translation along the x-axis.
  - dy: Translation along the y-axis.
  - dz: Translation along the z-axis.
  - theta: Rotation angle around the x-axis.
  - phi: Rotation angle around the y-axis.
  - psi: Rotation angle around the z-axis.
}
\description{
This function calculates default values for the parameters used in structure alignment.
The parameters include translation along x, y, z axes (dx, dy, dz), and rotation angles
(theta, phi, psi) based on the given 3D coordinates of two structures.
}
\examples{
\dontrun{
# Example: Get default values for structure alignment parameters
default_values <- get_default_values(coord1, coord2)
}

}
\references{
Borchers H (2022). pracma: Practical Numerical Math Functions. R package
version 2.4.2, \href{https://CRAN.R-project.org/package=pracma}{Link}.
Mathematics Stack Exchange: https://math.stackexchange.com/questions/180418/calculate-rotation-matrix-to-align-vector-a-to-vector-b-in-3d/897677#897677
}
