% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{get_rmsd}
\alias{get_rmsd}
\title{Get Root Mean Square Deviation (RMSD) from Protein Structure Alignment}
\usage{
get_rmsd(alignment)
}
\arguments{
\item{alignment}{List. Structure alignment results, including alignment
parameters, coordinates, and other information.
The list should contain the following elements:\cr
- N: Numeric. Number of common residues in the alignment.\cr
- coord1: Matrix. 3D coordinates of CA atoms for common residues in the
  first structure.\cr
- coord2: Matrix. 3D coordinates of CA atoms for common residues in the
  second structure.\cr
- values: Numeric vector. Alignment parameters.}
}
\value{
Numeric. Root Mean Square Deviation (RMSD) of the protein structure
  alignment.
}
\description{
This function calculates the Root Mean Square Deviation (RMSD) from the
alignment parameters and coordinates obtained in a structural alignment
between two protein structures.
}
\examples{
\dontrun{
# Example: Calculate RMSD from alignment
alignment_results <- get_alignment("structure1.pdb", "structure2.pdb",
                                  chain1 = 'A', chain2 = 'A',
                                  method = "alignment")
rmsd_value <- get_rmsd(alignment_results)
print(rmsd_value)
}

}
\seealso{
\code{\link{get_alignment}} for obtaining structural alignment between two
  protein structures.
\code{\link{optimize}} for optimizing alignment parameters.
\code{\link{estimate_d0}} for estimating initial distance parameters.
\code{\link{rmsd}} for calculating RMSD.
}
