% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{get_tmscore}
\alias{get_tmscore}
\title{Get TM Score from Protein Structure Alignment}
\usage{
get_tmscore(alignment)
}
\arguments{
\item{alignment}{List. Structure alignment results, including alignment
parameters, coordinates, and other information.
The list should contain the following elements: \cr
- N: Numeric. Number of common residues in the alignment. \cr
- coord1: Matrix. 3D coordinates of CA atoms for common residues in the
  first structure.\cr
- coord2: Matrix. 3D coordinates of CA atoms for common residues in the
  second structure.\cr
- values: Numeric vector. Alignment parameters.}
}
\value{
Numeric. TM score of the protein structure alignment.
}
\description{
This function calculates the TM score from the alignment parameters and
coordinates obtained in a structural alignment between two protein
structures.
}
\examples{
\dontrun{
# Example: Calculate TM score from alignment
alignment_results <- get_alignment("structure1.pdb", "structure2.pdb",
                                  chain1 = 'A', chain2 = 'A',
                                  method = "alignment")
optimized_alignment <- optimize(alignment_results)
tmscore <- get_tmscore(optimized_alignment)
print(tmscore)
}

}
\references{
Zhang, Y., and Skolnick, J. (2004). Scoring function for automated assessment
of protein structure template quality. \emph{Proteins, Structure, Function,
and Bioinformatics}, 57(4), 702–710.
\href{https://doi.org/10.1002/prot.20264}{Link}
}
\seealso{
\code{\link{get_alignment}} for obtaining alignment details.
\code{\link{optimize}} for optimizing alignment parameters.
\code{\link{estimate_d0}} for estimating initial distance parameters.
\code{\link{tm}} for calculating TM-score.
}
