% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{optimize}
\alias{optimize}
\title{Optimize Protein Structure Alignment}
\usage{
optimize(coord1, coord2, d02, values = NULL, restart = TRUE)
}
\arguments{
\item{coord1}{Numeric matrix. Coordinates of the reference protein structure.}

\item{coord2}{Numeric matrix. Coordinates of the target protein structure.}

\item{d02}{Numeric. Reference distance for the TM-score calculation.}

\item{values}{Numeric vector. Initial values for the optimization parameters.
If provided, the optimization continues from these values; otherwise, default values are used.}

\item{restart}{Logical. If TRUE, the optimization is restarted with default values; otherwise, it continues.}
}
\value{
A list containing the optimized values, TM-score, and RMSD.
  - values: Numeric vector. Optimized parameters for the alignment.
  - tmscore: Numeric. TM-score of the aligned structures.
  - rmsd: Numeric. Root Mean Square Deviation (RMSD) between the aligned structures.
}
\description{
This function performs optimization to improve the alignment of two protein structures based
on their atomic coordinates. The optimization aims to find the best parameters that minimize
the objective function, which includes the TM-score and RMSD calculations. The optimization
can be restarted with default values or continue from a given set of values.
}
\examples{
\dontrun{
# Example: Optimize protein structure alignment
reference_structure <- matrix(rnorm(300), ncol = 3)
target_structure <- matrix(rnorm(300), ncol = 3)
reference_d0 <- 10
optimization_results <- optimize(coord1 = reference_structure, coord2 = target_structure, d02 = reference_d0)
}

}
\references{
Byrd, R. H., Lu, P., Nocedal, J., & Zhu, C. (1995).
A limited memory algorithm for bound constrained optimization.
\emph{SIAM Journal on Scientific Computing}, 16(5), 1190–1208.
\href{https://doi.org/10.1137/091606}{Link}
}
\seealso{
\code{\link{get_default_values}} and \code{\link{get_current_values}} for parameter initialization.
\code{\link{tm}} and \code{\link{rmsd}} for TM-score and RMSD calculation.
\code{\link{stats::optim}} for optimization functions.
}
