% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTMscoreAlign.R
\name{runTMscoreAlign}
\alias{runTMscoreAlign}
\title{Launch Shiny App For Package TMscoreAlign}
\usage{
runTMscoreAlign()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
TODO: finish documentation for this
}
\examples{
\dontrun{
runTMscoreAlign()
}

}
\references{
TODO: finish documentation for this
}
\author{
Kevin Zhu, \email{kevqyzhu@gmail.com} # is this needed?
}
